[Unification] Move get_extended_attention_mask to utils
[ghstack-poisoned]
